test(integration): add integration tests to catch cyclic locking issue
see #194
